+++ lib/ui.dart
@@ -87,6 +87,8 @@ part 'src/ui/date_label.dart';
 part 'src/ui/value_box.dart';
 part 'src/ui/double_box.dart';
 part 'src/ui/int_box.dart';
+part 'src/ui/simple_check_box.dart';
+part 'src/ui/simple_radio_button.dart';
 
 part 'src/ui/inline_label.dart';
 part 'src/ui/inline_html.dart';